separate base filter, and introduce max line length limit
