+++ filter/tail/tail.go
@@ -5,12 +5,12 @@ import (
 	"io"
 
 	"github.com/koron-go/ringbuf"
-	"github.com/koron/nvgd/filter"
+	"github.com/koron/nvgd/internal/filterbase"
 )
 
 // Tail is "tail" like filter.
 type Tail struct {
-	filter.Base
+	filterbase.Base
 
 	rf bool
 	rb *ringbuf.Buffer[[]byte]
